Move getUploadedCpis into QueryTasksImpl as a private fun
